First commit to the project
Even though is the first commit to the project, there is already a lot
of code in the "main.py". Basically, the game is working for 2
players, and in each movement, the console output what happened in the
current game state. The game is very basic yet, but will have a lot of
improvements with it.