Inserting all the code, the first commit
Even though is the first commit to the project, there is already a lot
of code in it. The "main.py" file have basically everything that
is necessary to make a 1x1 game. For next commit, the work will be
focused in the AI, making possible to play with the computer.