New file and some organizations
In this commit, there is a lot of organization. First of all, it is
possible to found a new file, called "operational.py", which is now used
to make the basic game. This basic game is the 1x1 game, that is now
separated in just one file. Last but not least, the docstrings are
updated, and should explain more about the code.