Easy mode added
In this commit the biggest change to the project is the new easy mode.
Basically the "AI" selects two random numbers, which determine where
the AI must place the next "O". It is not really effective but for a
easy mode it's ok. It was still added some files and a lot of
docstrings. Most functions are used more than once,
making the code really reusable.