Added multiple modes
In this commit, there are a lot of changes in the possibilities to play.
First of all, now when the game lauches you must tipe: "start
<user/easy> <user/easy>, representing which character the player wants
to pick. The "easy" option is the "AI" one.. the computer picks aleatory
numbers to try to win the game. Several issues were also resolved.