Inseted medium AI level
Today there is a lot of progression on the AI, making a little smarter
option to play against. Actually, it's very hard to win against this AI.
Basically it has 3 steps. In the start, the "medium" AI look for
possible game winning moves. If there are no moves like that, it start
searching for blocks on the enemy. If has none of those options, it just
make a random move. Right now, the code is without comments in this part
and needs to be reorganized, but that will be in the next commit.