Inserted hard AI
New AI mode inserted. This AI uses minimax to work. Basically, minimax
is an algorithm that searches all possible moves for the AI and all
possibles moves to the other player, doing only safe moves. In this AI
level you can't win, but can draw a game. There is one problem yet, the
algorithm is taking too long to search in the first position (2s), but
this will be fixed in the next commit.