Some improvements has been made
This commit is just for organization and optimization. First all
docstrings are now updated. Now about optimization, two big things
happend in this commit. The next_play() function is a little bit longer
now but it is 1 sec faster than the last one, making this one better for
recursion. And now the first hard play is always the same, because it is
the best play to do in the first move.